5. Source Codes/recommender/requirements.txt
numpy
pandas
scipy
sklearn
gunicorn
flask
flask-restplus
scikit-surprise